+++ cmd/gox/install.sh
@@ -54,7 +54,60 @@ fi
 
 mv $appname $GOPATH/bin/
 
-# build wasm
+# build and install gdspx (PC platform shared library)
+echo "Building gdspx..."
+cd ../igox || exit
+
+go mod tidy
+
+# Get architecture
+GOARCH=$(go env GOARCH)
+
+if [ "$OS" = "Windows_NT" ]; then
+    # Windows: build DLL with proper naming
+    GOOS_NAME="windows"
+    EXT=".dll"
+    LIB_NAME="gdspx-${GOOS_NAME}-${GOARCH}${EXT}"
+    go build -ldflags="-checklinkname=0 -extldflags=-Wl,--allow-multiple-definition" -buildmode=c-shared -o "$LIB_NAME" .
+    if [ -f "$LIB_NAME" ]; then
+        cp -f "$LIB_NAME" "$GOPATH/bin/"
+        echo "Installed $LIB_NAME to $GOPATH/bin/"
+    else
+        echo "Error: $LIB_NAME build failed"
+        exit 1
+    fi
+elif [ "$(uname)" = "Darwin" ]; then
+    # macOS: build dylib with proper naming
+    GOOS_NAME="darwin"
+    EXT=".dylib"
+    LIB_NAME="gdspx-${GOOS_NAME}-${GOARCH}${EXT}"
+    go build -ldflags="-checklinkname=0" -buildmode=c-shared -o "$LIB_NAME" .
+    if [ -f "$LIB_NAME" ]; then
+        cp -f "$LIB_NAME" "$GOPATH/bin/"
+        echo "Installed $LIB_NAME to $GOPATH/bin/"
+    else
+        echo "Error: $LIB_NAME build failed"
+        exit 1
+    fi
+else
+    # Linux: build so with proper naming
+    GOOS_NAME="linux"
+    EXT=".so"
+    LIB_NAME="gdspx-${GOOS_NAME}-${GOARCH}${EXT}"
+    go build -ldflags="-checklinkname=0 -extldflags=-Wl,--allow-multiple-definition" -buildmode=c-shared -o "$LIB_NAME" .
+    if [ -f "$LIB_NAME" ]; then
+        cp -f "$LIB_NAME" "$GOPATH/bin/"
+        echo "Installed $LIB_NAME to $GOPATH/bin/"
+    else
+        echo "Error: $LIB_NAME build failed"
+        exit 1
+    fi
+fi
+
+
+cd ../gox || exit
+
+# build wasm (Web platform)
 if [ "$1" = "--web" ]; then
     go env -w GOFLAGS="-buildvcs=false"
     cd ../igox || exit
