+++ cmd/gox/main.go
@@ -48,5 +48,9 @@ func main() {
 	cmd.GoModTemplate = gomodtemplate
 
 	// Initialize the Args field if not already initialized
-	cmd.RunCmd(appName, appName, cmd.Version, projectFS, "template/project", "project")
+	err := cmd.RunCmd(appName, appName, cmd.Version, projectFS, "template/project", "project")
+	if err != nil {
+		println("failed to run cmd:", err)
+		return
+	}
 }
