+++ cmd/gox/pkg/command/run.go
@@ -134,3 +134,47 @@ func (pself *CmdTool) RunPureEngine(pargs ...string) error {
 func (pself *CmdTool) RunWithAiMode(pargs ...string) error {
 	return pself.RunPackMode(pargs...)
 }
+
+// RunInterpreted runs the project in interpreted mode.
+func (pself *CmdTool) RunInterpreted(pargs ...string) error {
+	// Get gdextension path from GOPATH/bin
+	extensionPath := path.Join(pself.GoBinPath, "runtime.gdextension")
+
+	// Verify runtime.gdextension exists
+	if _, err := os.Stat(extensionPath); os.IsNotExist(err) {
+		return fmt.Errorf("runtime.gdextension not found at %s. Please run 'spx install' first", extensionPath)
+	}
+
+	// Verify the shared library exists
+	GOOS := runtime.GOOS
+	GOARCH := runtime.GOARCH
+	var libExt string
+	switch GOOS {
+	case "windows":
+		libExt = ".dll"
+	case "darwin":
+		libExt = ".dylib"
+	default:
+		libExt = ".so"
+	}
+	libName := fmt.Sprintf("gdspx-%s-%s%s", GOOS, GOARCH, libExt)
+	libPath := path.Join(pself.GoBinPath, libName)
+	if _, err := os.Stat(libPath); os.IsNotExist(err) {
+		return fmt.Errorf("shared library %s not found at %s. Please run 'make install' first", libName, pself.GoBinPath)
+	}
+
+	// Build command arguments
+	args := []string{}
+	for i := 0; i < len(pargs); i++ {
+		if pargs[i] == "--path" {
+			i++ // Skip the path value
+			continue
+		}
+		args = append(args, pargs[i])
+	}
+	args = append(args, "--path", pself.RuntimeTempDir)
+	args = append(args, "--gdextpath", extensionPath)
+	args = append(args, "--no-header")
+	// Run the gdspxrt runtime
+	return util.RunCommandInDir(pself.RuntimeTempDir, pself.RuntimeCmdPath, args...)
+}
