+++ cmd/igox/embedpkg/embedpkg.go
@@ -1,0 +1,169 @@
+package embedpkg
+
+// All packages available in the ispx Wasm runtime.
+import (
+	// Embedded stdlib packages.
+	_ "github.com/goplus/ixgo/pkg/archive/tar"
+	_ "github.com/goplus/ixgo/pkg/archive/zip"
+	_ "github.com/goplus/ixgo/pkg/bufio"
+	_ "github.com/goplus/ixgo/pkg/bytes"
+	_ "github.com/goplus/ixgo/pkg/cmp"
+	_ "github.com/goplus/ixgo/pkg/compress/bzip2"
+	_ "github.com/goplus/ixgo/pkg/compress/flate"
+	_ "github.com/goplus/ixgo/pkg/compress/gzip"
+	_ "github.com/goplus/ixgo/pkg/compress/lzw"
+	_ "github.com/goplus/ixgo/pkg/compress/zlib"
+	_ "github.com/goplus/ixgo/pkg/context"
+	_ "github.com/goplus/ixgo/pkg/crypto"
+	_ "github.com/goplus/ixgo/pkg/crypto/aes"
+	_ "github.com/goplus/ixgo/pkg/crypto/cipher"
+	_ "github.com/goplus/ixgo/pkg/crypto/des"
+	_ "github.com/goplus/ixgo/pkg/crypto/dsa"
+	_ "github.com/goplus/ixgo/pkg/crypto/ecdh"
+	_ "github.com/goplus/ixgo/pkg/crypto/ecdsa"
+	_ "github.com/goplus/ixgo/pkg/crypto/ed25519"
+	_ "github.com/goplus/ixgo/pkg/crypto/elliptic"
+	_ "github.com/goplus/ixgo/pkg/crypto/hmac"
+	_ "github.com/goplus/ixgo/pkg/crypto/md5"
+	_ "github.com/goplus/ixgo/pkg/crypto/rand"
+	_ "github.com/goplus/ixgo/pkg/crypto/rc4"
+	_ "github.com/goplus/ixgo/pkg/crypto/rsa"
+	_ "github.com/goplus/ixgo/pkg/crypto/sha1"
+	_ "github.com/goplus/ixgo/pkg/crypto/sha256"
+	_ "github.com/goplus/ixgo/pkg/crypto/sha512"
+	_ "github.com/goplus/ixgo/pkg/crypto/subtle"
+	_ "github.com/goplus/ixgo/pkg/encoding"
+	_ "github.com/goplus/ixgo/pkg/encoding/asn1"
+	_ "github.com/goplus/ixgo/pkg/encoding/base32"
+	_ "github.com/goplus/ixgo/pkg/encoding/base64"
+	_ "github.com/goplus/ixgo/pkg/encoding/binary"
+	_ "github.com/goplus/ixgo/pkg/encoding/csv"
+	_ "github.com/goplus/ixgo/pkg/encoding/gob"
+	_ "github.com/goplus/ixgo/pkg/encoding/hex"
+	_ "github.com/goplus/ixgo/pkg/encoding/json"
+	_ "github.com/goplus/ixgo/pkg/encoding/pem"
+	_ "github.com/goplus/ixgo/pkg/encoding/xml"
+	_ "github.com/goplus/ixgo/pkg/errors"
+	_ "github.com/goplus/ixgo/pkg/fmt"
+	_ "github.com/goplus/ixgo/pkg/hash"
+	_ "github.com/goplus/ixgo/pkg/hash/adler32"
+	_ "github.com/goplus/ixgo/pkg/hash/crc32"
+	_ "github.com/goplus/ixgo/pkg/hash/crc64"
+	_ "github.com/goplus/ixgo/pkg/hash/fnv"
+	_ "github.com/goplus/ixgo/pkg/hash/maphash"
+	_ "github.com/goplus/ixgo/pkg/html"
+	_ "github.com/goplus/ixgo/pkg/html/template"
+	_ "github.com/goplus/ixgo/pkg/image"
+	_ "github.com/goplus/ixgo/pkg/image/color"
+	_ "github.com/goplus/ixgo/pkg/image/color/palette"
+	_ "github.com/goplus/ixgo/pkg/image/draw"
+	_ "github.com/goplus/ixgo/pkg/image/gif"
+	_ "github.com/goplus/ixgo/pkg/image/jpeg"
+	_ "github.com/goplus/ixgo/pkg/image/png"
+	_ "github.com/goplus/ixgo/pkg/io"
+	_ "github.com/goplus/ixgo/pkg/io/fs"
+	_ "github.com/goplus/ixgo/pkg/io/ioutil"
+	_ "github.com/goplus/ixgo/pkg/iter"
+	_ "github.com/goplus/ixgo/pkg/log"
+	_ "github.com/goplus/ixgo/pkg/log/slog"
+	_ "github.com/goplus/ixgo/pkg/maps"
+	_ "github.com/goplus/ixgo/pkg/math"
+	_ "github.com/goplus/ixgo/pkg/math/big"
+	_ "github.com/goplus/ixgo/pkg/math/bits"
+	_ "github.com/goplus/ixgo/pkg/math/cmplx"
+	_ "github.com/goplus/ixgo/pkg/math/rand"
+	_ "github.com/goplus/ixgo/pkg/math/rand/v2"
+	_ "github.com/goplus/ixgo/pkg/mime"
+	_ "github.com/goplus/ixgo/pkg/net/http"
+	_ "github.com/goplus/ixgo/pkg/net/netip"
+	_ "github.com/goplus/ixgo/pkg/net/url"
+	_ "github.com/goplus/ixgo/pkg/os"
+	_ "github.com/goplus/ixgo/pkg/path"
+	_ "github.com/goplus/ixgo/pkg/path/filepath"
+	_ "github.com/goplus/ixgo/pkg/reflect"
+	_ "github.com/goplus/ixgo/pkg/regexp"
+	_ "github.com/goplus/ixgo/pkg/regexp/syntax"
+	_ "github.com/goplus/ixgo/pkg/runtime"
+	_ "github.com/goplus/ixgo/pkg/slices"
+	_ "github.com/goplus/ixgo/pkg/sort"
+	_ "github.com/goplus/ixgo/pkg/strconv"
+	_ "github.com/goplus/ixgo/pkg/strings"
+	_ "github.com/goplus/ixgo/pkg/sync"
+	_ "github.com/goplus/ixgo/pkg/sync/atomic"
+	_ "github.com/goplus/ixgo/pkg/text/scanner"
+	_ "github.com/goplus/ixgo/pkg/text/tabwriter"
+	_ "github.com/goplus/ixgo/pkg/text/template"
+	_ "github.com/goplus/ixgo/pkg/text/template/parse"
+	_ "github.com/goplus/ixgo/pkg/time"
+	_ "github.com/goplus/ixgo/pkg/time/tzdata"
+	_ "github.com/goplus/ixgo/pkg/unicode"
+	_ "github.com/goplus/ixgo/pkg/unicode/utf16"
+	_ "github.com/goplus/ixgo/pkg/unicode/utf8"
+	_ "github.com/goplus/ixgo/pkg/unique"
+	// Excluded stdlib packages.
+	// _ "github.com/goplus/ixgo/pkg/container/heap"
+	// _ "github.com/goplus/ixgo/pkg/container/list"
+	// _ "github.com/goplus/ixgo/pkg/container/ring"
+	// _ "github.com/goplus/ixgo/pkg/crypto/tls"
+	// _ "github.com/goplus/ixgo/pkg/crypto/x509"
+	// _ "github.com/goplus/ixgo/pkg/crypto/x509/pkix"
+	// _ "github.com/goplus/ixgo/pkg/database/sql"
+	// _ "github.com/goplus/ixgo/pkg/database/sql/driver"
+	// _ "github.com/goplus/ixgo/pkg/debug/buildinfo"
+	// _ "github.com/goplus/ixgo/pkg/debug/dwarf"
+	// _ "github.com/goplus/ixgo/pkg/debug/elf"
+	// _ "github.com/goplus/ixgo/pkg/debug/gosym"
+	// _ "github.com/goplus/ixgo/pkg/debug/macho"
+	// _ "github.com/goplus/ixgo/pkg/debug/pe"
+	// _ "github.com/goplus/ixgo/pkg/debug/plan9obj"
+	// _ "github.com/goplus/ixgo/pkg/embed"
+	// _ "github.com/goplus/ixgo/pkg/encoding/ascii85"
+	// _ "github.com/goplus/ixgo/pkg/expvar"
+	// _ "github.com/goplus/ixgo/pkg/flag"
+	// _ "github.com/goplus/ixgo/pkg/go/ast"
+	// _ "github.com/goplus/ixgo/pkg/go/build"
+	// _ "github.com/goplus/ixgo/pkg/go/build/constraint"
+	// _ "github.com/goplus/ixgo/pkg/go/constant"
+	// _ "github.com/goplus/ixgo/pkg/go/doc"
+	// _ "github.com/goplus/ixgo/pkg/go/doc/comment"
+	// _ "github.com/goplus/ixgo/pkg/go/format"
+	// _ "github.com/goplus/ixgo/pkg/go/importer"
+	// _ "github.com/goplus/ixgo/pkg/go/parser"
+	// _ "github.com/goplus/ixgo/pkg/go/printer"
+	// _ "github.com/goplus/ixgo/pkg/go/scanner"
+	// _ "github.com/goplus/ixgo/pkg/go/token"
+	// _ "github.com/goplus/ixgo/pkg/go/types"
+	// _ "github.com/goplus/ixgo/pkg/go/version"
+	// _ "github.com/goplus/ixgo/pkg/index/suffixarray"
+	// _ "github.com/goplus/ixgo/pkg/log/syslog"
+	// _ "github.com/goplus/ixgo/pkg/mime/multipart"
+	// _ "github.com/goplus/ixgo/pkg/mime/quotedprintable"
+	// _ "github.com/goplus/ixgo/pkg/net"
+	// _ "github.com/goplus/ixgo/pkg/net/http/cgi"
+	// _ "github.com/goplus/ixgo/pkg/net/http/cookiejar"
+	// _ "github.com/goplus/ixgo/pkg/net/http/fcgi"
+	// _ "github.com/goplus/ixgo/pkg/net/http/httptest"
+	// _ "github.com/goplus/ixgo/pkg/net/http/httptrace"
+	// _ "github.com/goplus/ixgo/pkg/net/http/httputil"
+	// _ "github.com/goplus/ixgo/pkg/net/http/pprof"
+	// _ "github.com/goplus/ixgo/pkg/net/mail"
+	// _ "github.com/goplus/ixgo/pkg/net/rpc"
+	// _ "github.com/goplus/ixgo/pkg/net/rpc/jsonrpc"
+	// _ "github.com/goplus/ixgo/pkg/net/smtp"
+	// _ "github.com/goplus/ixgo/pkg/net/textproto"
+	// _ "github.com/goplus/ixgo/pkg/os/exec"
+	// _ "github.com/goplus/ixgo/pkg/os/signal"
+	// _ "github.com/goplus/ixgo/pkg/os/user"
+	// _ "github.com/goplus/ixgo/pkg/runtime/coverage"
+	// _ "github.com/goplus/ixgo/pkg/runtime/debug"
+	// _ "github.com/goplus/ixgo/pkg/runtime/metrics"
+	// _ "github.com/goplus/ixgo/pkg/runtime/pprof"
+	// _ "github.com/goplus/ixgo/pkg/runtime/trace"
+	// _ "github.com/goplus/ixgo/pkg/structs"
+	// _ "github.com/goplus/ixgo/pkg/syscall"
+	// _ "github.com/goplus/ixgo/pkg/testing"
+	// _ "github.com/goplus/ixgo/pkg/testing/fstest"
+	// _ "github.com/goplus/ixgo/pkg/testing/iotest"
+	// _ "github.com/goplus/ixgo/pkg/testing/quick"
+	// _ "github.com/goplus/ixgo/pkg/testing/slogtest"
+)
