+++ cmd/igox/main.go
@@ -1,5 +1,3 @@
-//go:build js && wasm
-
 package main
 
 //go:generate go tool qexp -outdir pkg github.com/goplus/spx/v2
@@ -10,15 +8,8 @@ package main
 import (
 	// Embedded third-party packages.
 	"github.com/goplus/spx/v2/cmd/igox/launcher"
-
-	_ "github.com/goplus/spx/v2/cmd/igox/pkg/github.com/goplus/spx/v2"
-	_ "github.com/goplus/spx/v2/cmd/igox/pkg/github.com/goplus/spx/v2/pkg/gdspx/pkg/engine"
 )
 
-type TestPlugin struct {
-	launcher.TestPlugin
-}
-
 func main() {
-	launcher.Run(launcher.Plugin{Name: "test", Plugin: &TestPlugin{}})
+	launcher.Run()
 }
